+++ docs/CHANGELOG.rst
@@ -6,6 +6,17 @@ All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
+==========
+Unreleased
+==========
+
+Added
+-----
+- Add ``upload_credentials`` endpoint
+- Mark endpoint ``upload_config_deprecated`` as deprecated and to be removed
+  after ``1. 1. 2025``
+
+
 ===================
 40.1.0 - 2024-07-09
 ===================